+++ lib/ueberauth_twitch.ex
@@ -1,0 +1,3 @@
+defmodule UeberauthTwitch do
+  @moduledoc false
+end