+++ lib/ueberauth_twitter.ex
@@ -1,3 +1,0 @@
-defmodule UeberauthTwitter do
-  @moduledoc false
-end