+++ CHANGELOG.md
@@ -27,4 +27,4 @@
 
 # v 0.1.0
 
-* Initial Twitter OAuth strategy
+* Initial Twitch OAuth strategy